+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This changelog keeps track of all changes to the Packs SDK. We follow convention
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed CLI execution of sync table metadata functions for sync tables with names that do not match their inner formula name.
+
 ## [1.12.0] - 2025-10-23
 
 ### Changed
@@ -996,5 +1000,4 @@ await myHelper(context);
 [1.10.2]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.10.2
 [1.11.1]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.11.1
 [1.11.0]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.11.0
-
 [1.12.0]: https://github.com/coda/packs-sdk/compare/v1.7.8...v1.12.0