+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This changelog keeps track of all changes to the Packs SDK. We follow convention
 
 ## [Unreleased]
 
+## 1.12.1-prerelease.1
+
+### Changed
+- Suggested Prompts support
+
 ## [1.12.1] - 2025-11-05
 
 ### Changed